Refactor get data code Relates #189
