+++ server/helpers/format-user-data.js
